requirements.txt
Flask
mysql-connector-python
tensorflow
keras
Pillow
